VT-6052: replace RestThreeDCartConfigBase with RestThreeDCartConfigV2
