+++ src/ThreeDCartAccess/RestApi/ThreeDCartServiceBase.cs
@@ -10,10 +10,10 @@ namespace ThreeDCartAccess.RestApi
 {
 	public abstract class ThreeDCartServiceBase
 	{
-		protected readonly RestThreeDCartConfigBase Config;
+		protected readonly RestThreeDCartConfigV2 Config;
 		internal readonly WebRequestServices WebRequestServices;
 
-		protected ThreeDCartServiceBase( RestThreeDCartConfigBase config )
+		protected ThreeDCartServiceBase( RestThreeDCartConfigV2 config )
 		{
 			Condition.Requires( config, "config" ).IsNotNull();
